test: update application from request tests
