+++ plugins/BEdita/API/tests/TestCase/Auth/EndpointAuthorizeTest.php
@@ -47,97 +47,6 @@ class EndpointAuthorizeTest extends TestCase
     ];
 
     /**
-     * Data provider for `testGetApplication` test case.
-     *
-     * @return array
-     */
-    public function getApplicationProvider()
-    {
-        return [
-            'standard' => [
-                1,
-                [
-                    'HTTP_X_API_KEY' => API_KEY,
-                ],
-            ],
-            'custom header' => [
-                1,
-                [
-                    'HTTP_X_CUSTOM_HEADER' => API_KEY,
-                ],
-                [],
-                [
-                    'apiKeyHeaderName' => 'X-Custom-Header',
-                ],
-            ],
-            'invalid API key' => [
-                new ForbiddenException('Invalid API key'),
-                [
-                    'HTTP_X_API_KEY' => 'this API key is invalid!',
-                ],
-            ],
-            'missing API key' => [
-                new ForbiddenException('Missing API key'),
-                [],
-                [],
-                [
-                    'blockAnonymousApps' => true,
-                ],
-            ],
-            'anonymous application' => [
-                null,
-                [],
-            ],
-            'query string api key' => [
-                1,
-                [],
-                [
-                    'api_key' => API_KEY,
-                ],
-            ],
-            'query string failure' => [
-                new ForbiddenException('Invalid API key'),
-                [],
-                [
-                    'api_key' => 'this API key is invalid!',
-                ]
-            ],
-        ];
-    }
-
-    /**
-     * Test getting application from request headers.
-     *
-     * @param int|\Exception $expected Expected application ID.
-     * @param array $environment Request headers.
-     * @param array $query Request query strings.
-     * @param array $config Configuration.
-     * @return void
-     *
-     * @dataProvider getApplicationProvider()
-     * @covers ::getApplication()
-     */
-    public function testGetApplication($expected, array $environment, array $query = [], array $config = [])
-    {
-        if ($expected instanceof \Exception) {
-            static::expectException(get_class($expected));
-            static::expectExceptionMessage($expected->getMessage());
-        }
-
-        CurrentApplication::getInstance()->set(null);
-        $authorize = new EndpointAuthorize(new ComponentRegistry(), $config);
-        $request = new ServerRequest(compact('environment', 'query'));
-
-        $authorize->authorize([], $request);
-
-        if ($expected === null) {
-            static::assertNull(CurrentApplication::getApplication());
-        } else {
-            static::assertEquals($expected, CurrentApplication::getApplication()->id);
-        }
-    }
-
-    /**
      * Data provider for `testGetEndpoint` test case.
      *
      * @return array
